chore(db): apply migrations 001-005 + regenerate types
- fix: cast NULL to integer in plan_features.banking_details insert
  (PG inferred text from bare NULL → 42804 type mismatch on apply).
- regenerate packages/types/database.types.ts from remote — picks up
  data_requests, ical_feeds, platform_staff_*, eft_banking_details,
  host_business_details, and the new subscription_history trigger.

Migrations applied to zlcivjgvtyeaszikqleu:
  001 banking_and_business_details
  002 create_platform_staff_rbac
  003 subscription_history_trigger
  004 data_requests
  005 ical_feeds

+++ supabase/migrations/20260525000001_banking_and_business_details.sql
@@ -99,12 +99,12 @@ CREATE POLICY business_owner_all ON public.host_business_details
 -- check_feature_permission so it can be disabled per-plan later without
 -- code changes.
 INSERT INTO public.plan_features (plan, feature_key, is_enabled, limit_value, description)
-SELECT DISTINCT plan, 'banking_details', true, NULL,
+SELECT DISTINCT plan, 'banking_details', true, NULL::integer,
        'Banking accounts + tax/business details for EFT and invoicing'
 FROM public.plan_features
 ON CONFLICT (plan, feature_key) DO UPDATE
   SET is_enabled = true,
-      limit_value = NULL;
+      limit_value = NULL::integer;
 
 -- ─── 6. Invoice trigger — capture banking + business ──────────
 CREATE OR REPLACE FUNCTION on_booking_confirmed_create_invoice()
